Move Schedduler instance under Configuration instance
